Move all Homepage API logic to the same file
